+++ apps/swirl-docs/src/pages/api/login.ts
@@ -1,6 +1,5 @@
 import { loginHandler } from "next-password-protect";
 
 export default loginHandler(process.env.SITE_PASSWORD!!, {
-  // Options go here (optional)
   cookieName: "next-password-protect",
 });