+++ apps/swirl-docs/types.d.ts
@@ -1,2 +1,1 @@
 declare module "remark-sectionize";
-declare module "balance-text";