+++ packages/swirl-components/src/components/swirl-popover/swirl-popover.tsx
@@ -49,6 +49,7 @@ export class SwirlPopover {
   @Prop() offset?: number | number[] = 8;
   @Prop() popoverId?: string;
   @Prop() placement?: Placement = "bottom-start";
+  @Prop() returnFocusToTrigger?: boolean = true;
   @Prop() trigger?: string | HTMLElement;
   @Prop() triggerContainer?: HTMLElement;
   @Prop() useContainerWidth?: boolean | string;
@@ -87,21 +88,29 @@ export class SwirlPopover {
     }
 
     const isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+    const isWKWebView = "webkit" in window;
+    const isSafariOrWKWebView = isSafari || isWKWebView;
+
     const target = event.target as HTMLElement;
     const relatedTarget = event.relatedTarget as HTMLElement;
     const activeElement = getActiveElement();
 
-    const popoverLostFocus =
-      !this.el.contains(target) &&
-      !this.el.contains(activeElement) &&
-      target !== this.triggerEl &&
-      !this.triggerEl?.contains(target) &&
-      (!isSafari ||
-        (isSafari &&
-          !this.el.contains(relatedTarget || target) &&
-          relatedTarget !== this.el));
+    // Check if the focus has moved outside the popover or its trigger.
+    const focusIsOutsidePopover =
+      !this.el.contains(target) && !this.el.contains(activeElement);
+    const focusIsNotOnTrigger =
+      target !== this.triggerEl && !this.triggerEl?.contains(target);
+    const extraCheckForSafariOrWKWebView =
+      isSafariOrWKWebView &&
+      !this.el.contains(relatedTarget || target) &&
+      relatedTarget !== this.el;
 
-    if (popoverLostFocus) {
+    // Close the popover if the focus is outside and additional checks for Safari or WKWebView pass.
+    if (
+      focusIsOutsidePopover &&
+      focusIsNotOnTrigger &&
+      (!isSafariOrWKWebView || extraCheckForSafariOrWKWebView)
+    ) {
       this.close();
     }
   }
@@ -156,7 +165,10 @@ export class SwirlPopover {
     }, 150);
 
     this.unlockBodyScroll();
-    this.getNativeTriggerElement()?.focus();
+
+    if (this.returnFocusToTrigger) {
+      this.getNativeTriggerElement()?.focus();
+    }
   }
 
   /**