+++ packages/swirl-components/src/components/swirl-search/swirl-search.stories.ts
@@ -15,10 +15,6 @@ export default {
 const Template = (args) => {
   const element = generateStoryElement("swirl-search", args);
 
-  element.addEventListener("inputInput", (event) => {
-    console.log(event);
-  });
-
   return element;
 };