+++ packages/swirl-components/src/components/swirl-theme-provider/swirl-theme-provider.stories.ts
@@ -89,10 +89,6 @@ const Template = (args) => {
 
   element.append(lightModeButton, darkModeButton, tenantButton, resetButton);
 
-  element.addEventListener("themeChange", (event: CustomEvent) => {
-    console.log(event.detail);
-  });
-
   return element;
 };