+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Tue, 24 Oct 2023 12:50:31 GMT
+// Generated on Wed, 15 Nov 2023 11:53:17 GMT